// src/services/api/models/block-chain-request.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Conifer.NFT.API
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { Blockchain } from './blockchain';
/**
 * 
 * @export
 * @interface BlockChainRequest
 */
export interface BlockChainRequest {
    /**
     * 
     * @type {Blockchain}
     * @memberof BlockChainRequest
     */
    blockchain?: Blockchain;
    /**
     * 
     * @type {string}
     * @memberof BlockChainRequest
     */
    ownerAddress?: string | null;
    /**
     * 
     * @type {number}
     * @memberof BlockChainRequest
     */
    limit?: number;
    /**
     * 
     * @type {number}
     * @memberof BlockChainRequest
     */
    offset?: number;
    /**
     * 
     * @type {string}
     * @memberof BlockChainRequest
     */
    searchValue?: string | null;
}